Updated header file

+++ function_pointers/1-array_iterator.c
@@ -9,7 +9,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	size_t i;
 
 	if (array != NULL)
 	{
